process count overlaps only if file doesn't exist

+++ create_count_overlaps.R
@@ -6,13 +6,6 @@ library(dplyr)
 library(tibble)                                                                 
 library(exomeCopy)
 
-args = commandArgs(trailingOnly=TRUE)
-if (length(args) != 1) {
-  stop("One argument must be supplied", call.=FALSE)
-} else {
-  CELL_NUMBER_FILTER = as.integer(args[1])
-} 
-
 get_sample_name <- function(file) {
   if (grepl("frontal_cortex", file)) {
     sample_name = "frontal_cortex"
@@ -52,7 +45,7 @@ get_sample_cell_types <- function(sample, sample_name, metadata) {
   sample_cellIDs = filtered_metadata[["cellID"]]                  
   sample_barcodes_in_metadata = unlist(lapply(sample_cellIDs, 
                                               function(x) unlist(strsplit(x, 
-                                                                          '\\+'))[2]))
+                                                                    '\\+'))[2]))
   sample = sample[sample$name %in% sample_barcodes_in_metadata]   
   sample_idx_in_metadata = match(sample$name, sample_barcodes_in_metadata)
   sample$cell_type = filtered_metadata[unlist(sample_idx_in_metadata), 
@@ -60,7 +53,14 @@ get_sample_cell_types <- function(sample, sample_name, metadata) {
   return(as.tibble(sample))
 }
 
-#load('/ahg/regevdata/projects/ICA_Lung/Bruno/celloforigin_prj/hg19.1Mb.ranges.Polak.Nature2015.RData')
+
+args = commandArgs(trailingOnly=TRUE)
+if (length(args) != 1) {
+  stop("One argument must be supplied", call.=FALSE)
+} else {
+  CELL_NUMBER_FILTER = as.integer(args[1])
+} 
+
 load('raw_dir/mutation_data/hg19.1Mb.ranges.Polak.Nature2015.RData')
 
 interval.ranges
@@ -77,20 +77,25 @@ for (file in setdiff(list.files("raw_dir/bed_files/"),
                    "count_overlaps", paste(file_path_sans_ext(file, TRUE),
                                            "rds", sep="."), sep="_")
   subdivided_filename = paste("count_filter", CELL_NUMBER_FILTER,  
-                              "subdivided_count_overlaps", paste(file_path_sans_ext(file, TRUE),
-                                                                 "rds", sep="."), sep="_")
+                              "subdivided_count_overlaps", 
+                              paste(file_path_sans_ext(file, TRUE),
+                                    "rds", sep="."), sep="_")
   if (!file.exists(filename) || !file.exists(subdivided_filename)) {
     print(paste("Processing", file, sep= " "))
     sample = import(paste("raw_dir", "bed_files", file, sep="/"), format="bed")
     sample_name = get_sample_name(file)
     sample <- get_sample_cell_types(sample, sample_name, metadata)
     sample <- filter_sample_by_cell_number(sample, CELL_NUMBER_FILTER)
-    count_overlaps <- get_count_overlaps(sample, interval.ranges)
-    count_overlaps_subdivided <- get_count_overlaps(sample, 
-                                                    subdivided_interval_ranges)
-    saveRDS(count_overlaps, paste("count_overlap_data", filename, sep="/"))
-    saveRDS(count_overlaps_subdivided, paste("count_overlap_data", 
-                                             subdivided_filename, 
-                                             sep="/"))
+    if (!file.exists(filename)) {
+      count_overlaps <- get_count_overlaps(sample, interval.ranges)
+      saveRDS(count_overlaps, paste("count_overlap_data", filename, sep="/"))
+    }
+    if (!file.exists(subdivided_filename)) {
+      count_overlaps_subdivided <- get_count_overlaps(sample, 
+                                                     subdivided_interval_ranges)
+      saveRDS(count_overlaps_subdivided, paste("count_overlap_data", 
+                                               subdivided_filename, 
+                                               sep="/"))
+    }
   }
 }
